+++ src/Ext/Techno/Hooks.Transport.cpp
@@ -1,7 +1,6 @@
 #include "Body.h"
 
-#include <Ext/House/Body.h>
-
+#include <Ext/Scenario/Body.h>
 
 DEFINE_HOOK_AGAIN(0x6FA33C, TechnoClass_ThreatEvals_OpenToppedOwner, 0x6) // TechnoClass::AI
 DEFINE_HOOK_AGAIN(0x6F89F4, TechnoClass_ThreatEvals_OpenToppedOwner, 0x6) // TechnoClass::EvaluateCell
@@ -84,10 +83,10 @@ DEFINE_HOOK(0x71067B, TechnoClass_EnterTransport, 0x7)
 		if (pTransTypeExt->Passengers_SyncOwner && pTransTypeExt->Passengers_SyncOwner_RevertOnExit)
 			pExt->OriginalPassengerOwner = pPassenger->Owner;
 
-		if (pPassenger->WhatAmI() != AbstractType::Aircraft && pType->Ammo > 0 && pExt->TypeExtData->ReloadInTransport)
+		if (pPassenger->WhatAmI() != AbstractType::Aircraft && pPassenger->WhatAmI() != AbstractType::Building
+			&& pType->Ammo > 0 && pExt->TypeExtData->ReloadInTransport)
 		{
-			auto const pOwnerExt = HouseExt::ExtMap.Find(pThis->Owner);
-			pOwnerExt->OwnedTransportReloaders.push_back(pExt);
+			ScenarioExt::Global()->TransportReloaders.push_back(pExt);
 		}
 	}
 
@@ -105,17 +104,18 @@ DEFINE_HOOK(0x4DE722, FootClass_LeaveTransport, 0x6)
 		auto const pExt = TechnoExt::ExtMap.Find(pPassenger);
 		auto const pTransTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
 
+		// Remove from transport reloader list before switching house
+		if (pPassenger->WhatAmI() != AbstractType::Aircraft && pPassenger->WhatAmI() != AbstractType::Building
+			&& pType->Ammo > 0 && pExt->TypeExtData->ReloadInTransport)
+		{
+			auto& vec = ScenarioExt::Global()->TransportReloaders;
+			vec.erase(std::remove(vec.begin(), vec.end(), pExt), vec.end());
+		}
+
 		if (pTransTypeExt->Passengers_SyncOwner && pTransTypeExt->Passengers_SyncOwner_RevertOnExit &&
 			pExt->OriginalPassengerOwner)
 		{
 			pPassenger->SetOwningHouse(pExt->OriginalPassengerOwner, false);
-		}
-
-		if (pThis->WhatAmI() != AbstractType::Aircraft && pType->Ammo > 0 && pExt->TypeExtData->ReloadInTransport)
-		{
-			auto const pOwnerExt = HouseExt::ExtMap.Find(pThis->Owner);
-			auto& vec = pOwnerExt->OwnedTransportReloaders;
-			vec.erase(std::remove(vec.begin(), vec.end(), pExt), vec.end());
 		}
 	}
